% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitterProfile.R
\name{hitterProfile}
\alias{hitterProfile}
\title{Create a scaled hitter profile from user defined offensive metrics}
\usage{
hitterProfile(
  player = NULL,
  wrc_plus = NULL,
  ops = NULL,
  obp = NULL,
  slg = NULL,
  avg = NULL,
  iso = NULL,
  war = NULL,
  woba = NULL,
  bb_pct = NULL,
  barrel_pct = NULL,
  hard_hit_pct = NULL
)
}
\arguments{
\item{player}{The name of the player.}

\item{wrc_plus}{Weighted runs created plus.}

\item{ops}{On-base plus slugging.}

\item{obp}{On-base percentage.}

\item{slg}{Slugging percentage.}

\item{avg}{Batting average.}

\item{iso}{Isolated power.}

\item{war}{Wins above replacement.}

\item{woba}{Weighted on-base average.}

\item{bb_pct}{Walk percentage.}

\item{barrel_pct}{Barrel percentage.}

\item{hard_hit_pct}{Hard-hit percentage.}
}
\value{
A data frame with the player name, selected metrics, raw inputted values,
and scaled values.
}
\description{
This function creates a hitter profile from selected offensive baseball metrics.
The user can enter any combination of the offensive metrics listed, which will return
a data frame with the raw inputted values and scaled values. The scaled values can
then be inputted into the \code{hitterRadar} function, that will plot the metrics on
a radar plot.
}
\examples{
judge_profile <- hitterProfile(
  player = "Aaron Judge",
  wrc_plus = 175,
  ops = 1.110,
  obp = 0.458,
  slg = 0.701,
  avg = 0.322,
  iso = 0.379
)

judge_profile
}
