% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitterRadar.R
\name{hitterRadar}
\alias{hitterRadar}
\title{Create a radar plot from a hitter profile}
\usage{
hitterRadar(profile_data)
}
\arguments{
\item{profile_data}{A data frame created by \code{hitter_profile()}, containing
a \code{metric} columns, a \code{scaled_value} column, and a \code{player} column.}
}
\value{
A radar plot displaying the chosen hitter's scaled offensive production
}
\description{
This function creates a radar plot using the scaled values from a hitter profile
data frame. It is desgined to work with the output from \code{hitterProfile()}.
}
\examples{
judge_profile <- hitterProfile(
  player = "Aaron Judge",
  ops = 1.000,
  avg = 0.300,
  slg = 0.700,
  wrc_plus = 175,
  iso = 0.210,
  war = 9.5
)

hitterRadar(judge_profile)
}
